+++ rrt_star.hpp
@@ -7,7 +7,7 @@
 
 class RRT_Star{
     public:
-        RRT_Star(State& s_init, State& s_goal, cv::Mat& image, State& origin, double& resolution, Segments& obstacles_segments){
+        RRT_Star(State& s_init, State& s_goal, cv::Mat& image, State& origin, long double& resolution, Segments& obstacles_segments){
             this->s_init = s_init;
             this->s_goal = s_goal;
             this->image = image;
@@ -26,16 +26,16 @@ class RRT_Star{
         State s_init; /// Initial state.
         State s_goal; /// Goal state.
         int iterations; /// Number of states to generate.
-        double time_step; /// Time step.
-        double linear_velocity; /// Linear velocity.
+        long double time_step; /// Time step.
+        long double linear_velocity; /// Linear velocity.
         nlohmann::json angular_velocity; /// Angular velocity.
-        double search_radius; /// Search radius.
-        double goal_threshold; /// Goal threshold.
-        double max_steering_angle; /// Max Steering Angle.
+        long double search_radius; /// Search radius.
+        long double goal_threshold; /// Goal threshold.
+        long double max_steering_angle; /// Max Steering Angle.
         /// ---
         cv::Mat image; /// Image is an object that stores the matrix of the loaded image.
         State origin; /// Origin is the coordinate of the bottom-left corner that corresponds to the pixel location (0,0) of the matrix.
-        double resolution; /// Resolution refers to the size of each cell or pixel in the occupancy grid map represented by a PGM file, expressed in meters per pixel.
+        long double resolution; /// Resolution refers to the size of each cell or pixel in the occupancy grid map represented by a PGM file, expressed in meters per pixel.
         int width; /// Width of the map in pixels.
         int height; /// Height of the map in pixels.
         Segments obstacles_segments; /// Collection of segments of polygons composing obstacles.
@@ -56,17 +56,17 @@ class RRT_Star{
         void rewire_tree(State*, std::vector<State*>&); /// Rewire tree.
         Path shortest_path(void); /// Shortest path.
         // ---
-        double calculate_euclidean_distance(double&, double&, double&, double&); /// Calculate euclidean disntace.
-        double calculate_distance(double&); /// Calculate distance from linear velocity.
-        double calculate_angle(double&); /// Caclulate distance from angular velocity.
-        double calculate_cost(State&, State&); /// Calculate cost.
+        long double calculate_euclidean_distance(long double&, long double&, long double&, long double&); /// Calculate euclidean disntace.
+        long double calculate_distance(long double&); /// Calculate distance from linear velocity.
+        long double calculate_angle(long double&); /// Caclulate distance from angular velocity.
+        long double calculate_cost(State&, State&); /// Calculate cost.
         Velocity inverse_odometry(State&, State&); /// Calculate velocities.
         bool do_intersect(cv::Point&, cv::Point&, cv::Point&, cv::Point&); /// Verify the intersection of a segment with obstacles.
         int orientation(cv::Point&, cv::Point&, cv::Point&); /// Returns the orientation of a triplet.
         bool on_segment(cv::Point&, cv::Point&, cv::Point&); /// Returns the collinearity of a triplet.
-        void swap(double*, double*); /// Swaps two elements of an array.
-        int x_convert_to_pixel(double&); /// Convert x-axis real coordinate to map coordinate
-        int y_convert_to_pixel(double&); /// Convert y-axis real coordinate to map coordinate
-        double get_random(double&); /// Takes a negative double and return a random number between [-arg, arg].
+        void swap(long double*, long double*); /// Swaps two elements of an array.
+        int x_convert_to_pixel(long double&); /// Convert x-axis real coordinate to map coordinate
+        int y_convert_to_pixel(long double&); /// Convert y-axis real coordinate to map coordinate
+        long double get_random(long double&); /// Takes a negative double and return a random number between [-arg, arg].
 };
 #endif
